consumer/requirements.txt
numpy==1.16.4
kafka-python
opencv_python==3.4.3.18
python-dotenv==0.10.3
Click==7.0
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
opencv-contrib-python==4.1.0.25
Pillow==6.1.0
pymongo==3.8.0
Werkzeug==0.15.5
absl-py==0.7.1
appnope==0.1.0
astor==0.8.0
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
Jinja2==2.10.1
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
networkx==2.3
pandas==0.25.0
parso==0.5.1
pexpect==4.7.0
phue==1.1
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
protobuf==3.9.0
ptyprocess==0.6.0
pygame==1.9.6
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
soco==0.17
tb-nightly==1.14.0a20190603
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
Theano==1.0.4
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.5
wget==3.2
wrapt==1.11.2
xmltodict==0.12.0
